Added mute for hush hush deving - should be environment based but I cba

+++ app/jarfis.js
@@ -8,43 +8,46 @@ const token = process.env.TOKEN;
 const greetings = ['hey', 'hi', 'yo', 'sup', 'sound', 'safe', 'whaddup'];
 const prefix = '!'; //this is still hard coded in commands.js
 const insults = fs.readFileSync('app/data/insults.txt').toString().split('\n');
+const mute = false; //stops bot responses
 
 client.on('ready', () => {
   console.log('ready to meme');
   if (process.env.ENV ===`Kanto`) {insultRand();}
 });
 
-client.on('message', msg => {
-  //loop through the commands module if msg starts with prefix
-  if (msg.content.startsWith(prefix)) {
-    args = msg.content.slice(prefix.length).split(' ');
+if (mute === false) {
+  client.on('message', msg => {
+    //loop through the commands module if msg starts with prefix
+    if (msg.content.startsWith(prefix)) {
+      args = msg.content.slice(prefix.length).split(' ');
 
-    if (args[0] in commands) {
-      commands[args[0]].execute(client, msg, args);
+      if (args[0] in commands) {
+        commands[args[0]].execute(client, msg, args);
+      }
     }
-  }
 
-  //these still need doing
-  if (msg.content == `bye ${getBot(msg)}`) {
-    msg.channel.send('see ya, wouldn\'t wanna meme ya');
-  }
+    //these still need doing
+    if (msg.content == `bye ${getBot(msg)}`) {
+      msg.channel.send('see ya, wouldn\'t wanna meme ya');
+    }
 
-  if (greetings.indexOf(msg.content.split(' ')[0]) > -1 && msg.content.split(' ')[1] == getBot(msg)) {
-    msg.channel.send('hi bitches');
-  }
+    if (greetings.indexOf(msg.content.split(' ')[0]) > -1 && msg.content.split(' ')[1] == getBot(msg)) {
+      msg.channel.send('hi bitches');
+    }
 
-  if (msg.content == 'too pure for this world') {
-    msg.channel.send(`thank you senpai`);
-  }
+    if (msg.content == 'too pure for this world') {
+      msg.channel.send(`thank you senpai`);
+    }
 
-  if (msg.content == '/lenny') {
-    msg.delete()
-      .then()
-      .catch(console.error);
-      
-    msg.channel.send(`( ͡° ͜ʖ ͡°)`);
-  }
-});
+    if (msg.content == '/lenny') {
+      msg.delete()
+        .then()
+        .catch(console.error);
+        
+      msg.channel.send(`( ͡° ͜ʖ ͡°)`);
+    }
+  });
+}
 
 function insultRand() {
   var users = client.users.array(),
